change i8 id to i32

+++ src/trash_manager.rs
@@ -155,7 +155,7 @@ pub fn add_all_elements_to_trash(
     }
 }
 
-pub fn remove_all_elements(connection: &Connection, is_test: bool, trash_items_ids: &[i8]) {
+pub fn remove_all_elements(connection: &Connection, is_test: bool, trash_items_ids: &[i32]) {
     trash_items_ids.iter().for_each(|trash_item_id| {
         let trash_item = data_manager::find_trash_item_by_id(connection, is_test, *trash_item_id);
         remove_element(&trash_item, is_test);
@@ -182,7 +182,7 @@ fn remove_element(trash_item: &TrashItem, is_test: bool) {
     );
 }
 
-pub fn restore_all_elements(connection: &Connection, is_test: bool, trash_items_ids: &[i8]) {
+pub fn restore_all_elements(connection: &Connection, is_test: bool, trash_items_ids: &[i32]) {
     trash_items_ids.iter().for_each(|trash_item_id| {
         let trash_item = data_manager::find_trash_item_by_id(connection, is_test, *trash_item_id);
         restore_element(&trash_item, is_test);
